requirements.txt
djangorestframework
django-redis
mysqlclient
django-celery-results